+++ app/(dashboard)/mydashboard/_components/invitations.tsx
@@ -1,10 +1,8 @@
 import { INITIAL_NUMBER_OF_USERS } from '@/constants/TEAM_BASE_URL';
-import search from '@/public/icons/search.svg';
 import { InvitationResponse } from '@/types/invitations';
-import Image from 'next/image';
 
 import { getInvitations } from '../actions';
-import InvitationList from './invitation-list';
+import InvitationContainer from './invitation-container';
 
 export default async function Invitations() {
   // NOTE - inviteAccepted null인 경우만 조회됨
@@ -18,17 +16,7 @@ export default async function Invitations() {
       <h2 className="mb-5 text-xl font-bold text-gray-700">
         초대받은 대시보드
       </h2>
-      <div className="relative">
-        <input
-          type="text"
-          className="h-9 w-full rounded-md border border-gray-300 py-[10px] pl-[44px] placeholder:text-sm placeholder:text-gray-400"
-          placeholder="검색"
-        />
-        <div className="absolute left-[20px] top-[50%] translate-y-[-50%]">
-          <Image src={search} width={16} height={16} alt="검색" />
-        </div>
-      </div>
-      <InvitationList
+      <InvitationContainer
         initialInvitations={invitations}
         initialCursorId={cursorId}
       />
